# api/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: INDIS API
  description: |
    # INDIS — Iran National Digital Identity System

    Gateway REST API for mobile clients (Android / iOS / HarmonyOS), the React PWA,
    verifier terminals, and the government portal.

    ## Authentication

    All protected routes require one of:
    - **Bearer JWT** — `Authorization: Bearer <token>` — HS256-signed JWT with claims
      `sub` (citizen DID), `ministry`, `role`, `exp`.
    - **X-API-Key** — `X-API-Key: <key>` — service-to-service key whose SHA-256 hash is
      registered in the gateway's `API_KEYS` configuration.

    Public routes that skip authentication:
    - `GET /health`
    - `GET /v1/identity/{did}`
    - `GET /v1/credential/{id}`
    - `POST /v1/electoral/verify`
    - `POST /v1/verifier/verify`
    - All `/v1/ussd` routes

    ## Privacy by Architecture

    Verifiers receive **only boolean ZK-proof results** — never raw citizen attributes.
    All verification events are appended to the tamper-evident audit log (PRD §FR-007).

    ## Localisation

    Dates accept and return RFC 3339 strings. The mobile client is responsible for
    converting to Solar Hijri (Persian calendar) using `pkg/i18n`.

    ## Rate Limiting

    100 requests per second per IP by default (configurable via `RATE_LIMIT_RPS`).
    Excess requests receive HTTP 429.
  version: "1.0.0"
  contact:
    name: INDIS Technical Team
    email: tech@indis.ir
  license:
    name: AGPL-3.0
servers:
...
    description: Staging
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256-signed JWT. Claims: `sub` (citizen DID, e.g. `did:indis:abc123`),
        `ministry` (e.g. `interior`), `role` (e.g. `operator`), `exp` (Unix timestamp).
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Service-to-service API key. The SHA-256 hash of this value must be registered
        in the gateway's `API_KEYS` environment variable.
  schemas:
...
    AppendEventRequest:
      type: object
      required: [action, category]
      properties:
        category:
          type: integer
          description: |
            EventCategory enum value:
            0=UNSPECIFIED, 1=IDENTITY, 2=CREDENTIAL, 3=ENROLLMENT,
            4=BIOMETRIC, 5=ELECTORAL, 6=JUSTICE, 7=ADMIN, 8=SYSTEM
        action:
          type: string
          description: Action string e.g. "did.register", "credential.revoke"
          example: "credential.revoke"
        actor_did:
          type: string
          description: DID of the operator; empty for system-generated events
        subject_did:
          type: string
          description: DID of the affected citizen; empty if not applicable
        resource_id:
          type: string
          description: Credential ID, enrollment ID, etc.
        service_id:
          type: string
          description: Name of the originating service
        metadata:
          type: string
          format: byte
          description: JSON metadata — no PII (PRD §FR-007)
    AppendEventResponse:
      type: object
      properties:
        event_id:
          type: string
        prev_hash:
          type: string
          description: SHA-256 hex of the previous log entry (hash chain integrity)
        timestamp:
          type: string
          format: date-time
    AuditEvent:
      type: object
      properties:
        event_id:
          type: string
        category:
          type: integer
        action:
          type: string
        actor_did:
          type: string
        subject_did:
          type: string
        resource_id:
          type: string
        service_id:
          type: string
        prev_hash:
          type: string
        entry_hash:
          type: string
        timestamp:
          type: string
          format: date-time
    QueryEventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        next_page_token:
          type: string

    # ── Notification ──────────────────────────────────────────────────────────
...
paths:
...
  /v1/audit/events:
    post:
      summary: Append an audit event
      description: |
        Appends an immutable event to the tamper-evident audit log. Each event is
        hash-chained to the previous entry (PRD §FR-007). Requires API key — internal
        use only; not accessible to citizen-facing clients.
      operationId: appendAuditEvent
      tags: [Audit]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppendEventRequest"
      responses:
        "201":
          description: Event appended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppendEventResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    get:
      summary: Query audit events
      description: |
        Returns audit events matching the given filters. Requires ministry role.
        Results are paginated using `next_page_token`.
      operationId: queryAuditEvents
      tags: [Audit]
      parameters:
        - name: actor_did
          in: query
          schema:
            type: string
          description: Filter by actor DID
        - name: subject_did
          in: query
          schema:
            type: string
          description: Filter by subject (citizen) DID
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start of time range (RFC 3339)
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End of time range (RFC 3339)
        - name: page_token
          in: query
          schema:
            type: string
          description: Pagination cursor
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryEventsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"

  # ── Card ──────────────────────────────────────────────────────────────────────
  /v1/card/{did}:
    get:
      summary: Retrieve a digital identity card
      description: Returns the digital identity card for the given citizen DID.
      operationId: getCard
      tags: [Card]
      parameters:
        - name: did
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Digital identity card
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/card/generate:
    post:
      summary: Generate a digital identity card
      description: |
        Generates or regenerates the digital identity card for the authenticated citizen.
        The card includes a QR code encoding a ZK-proof presentation request.
      operationId: generateCard
      tags: [Card]
      responses:
        "200":
          description: Card generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  card_id:
                    type: string
                  qr_code:
                    type: string
                    format: byte
                    description: Base64-encoded QR code PNG
                  expires_at:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
...